Think this is how i should upload the stuff i have worked on, not completely sure.
This is a program that finds the eigenvalues based on admittance transfer functions. ATM only created analytical transfer functions for UIC (a specific type of VSC control/design).

There seems to be some issues still at certain parameters when multiple units are present in the system, but I will debug till I fall over.

Lots of opertunities here, I think the code is fairly easy to understand, given the circumstances. The core of the code I have created is in eigFromY.